Split up main overload of abc::map::key_lookup()
Now divided in lookup_key(), lookup_key_or_find_empty_bucket() and
find_empty_bucket(). This is in reaction to a bug found in key_lookup()
that was caused by the 3 different match criteria being mashed into a
single if statement; making the if more robust would probably make it too
slow.