Remove unusable type_void_adapter::equal and related code
Equality comparison is typically provided by a functor that only the most
derived class of a container, being a template, has access to, and as a
member method, not a static function.
This means that either type_void_adapter be changed to store a this
pointer, or that two function calls be used (tva.equal -> static wrapper
-> operator() ).

A better alternative is to follow the pattern used by map/map_impl, which
involves only the static wrapper for operator().